Metaverse/ms-dataquality-p/lambda-docker-google-drive/requirements.txt
google-auth
google-auth-oauthlib
google-api-python-client
boto3